細かい修正

+++ src/diffdata.cpp
@@ -36,7 +36,7 @@ namespace schrac {
         mo_.reserve(osize);
         mi_.reserve(isize);
 
-        auto const len = grid_num - isize;
+        auto const len = grid_num - boost::numeric_cast<std::int32_t>(isize);
 
         for (auto i = 0; i <= mp_o_; i++) {
             auto const x = pdata_->xmin_ + static_cast<double>(i) * dx_;
